Move some files around, giving temp name due to conflict
