+++ src/main/java/rs117/hd/data/materials/NewGroundMaterial.java
@@ -1,21 +1,23 @@
-package rs117.hd.data.materials.groundMaterial;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.extern.slf4j.Slf4j;
+package rs117.hd.data.materials;
 
 import java.util.Random;
-import rs117.hd.data.materials.Material;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
 @Getter
 @Slf4j
-@AllArgsConstructor
-public class GroundMaterial {
+@NoArgsConstructor // Called by GSON when parsing JSON
+public class NewGroundMaterial {
+	public static final NewGroundMaterial NONE = new NewGroundMaterial("NONE", Material.NONE);
 
-	public static final GroundMaterial NONE = new GroundMaterial("NONE", new Material[] { Material.NONE });
+	public String name;
+	public Material[] materials;
 
-	private String name;
-	private Material[] materials;
+	public NewGroundMaterial(String name, Material... materials) {
+		this.name = name;
+		this.materials = materials;
+	}
 
 	/**
 	 * Get a random material based on given coordinates.
@@ -29,7 +31,7 @@ public class GroundMaterial {
 		// Generate a seed from the tile coordinates for consistent 'random' results between scene loads.
 		long seed = (long) (plane + 1) * 10 * (worldX % 100) * 20 * (worldY % 100) * 30;
 		Random randomTex = new Random(seed);
-		int randomIndex = randomTex.nextInt(this.materials.length);
-		return this.materials[randomIndex];
+		int randomIndex = randomTex.nextInt(materials.length);
+		return materials[randomIndex];
 	}
 }
