+++ src/main/java/rs117/hd/scene/tile_overrides/TileOverride.java
@@ -13,7 +13,6 @@ import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import rs117.hd.data.WaterType;
-import rs117.hd.data.materials.groundMaterial.GroundMaterial;
 import rs117.hd.scene.AreaManager;
 import rs117.hd.scene.areas.Area;
 import rs117.hd.utils.Props;
